requirement.txt
flask

flask-Login

werkzeug


pypaystack


Pillow


google-generativeai

rep

ortlab


google-api-core


tqdm